refactor: 카드 등록 화면 패키지 이동
- 카드 등록 화면 관련 파일을 `ui.payments`에서 `ui.screen.registration` 패키지로 이동합니다.
